refactor:  코드 리펙토링
리펙토링:
파일: main.py,
폴더: common -> core
기능 분리: main.py(auth) -> auth.py
