+++ .gitignore
@@ -1,2 +1,3 @@
 .DS_Store
 app/static/pictures/*
+app/core/consts.py
