--- app/common/consts.py
@@ -1,2 +1,0 @@
-
-DB_URL = 'sqlite:///../global_chat.db' # sqlite
