+++ app/database/crud.py
@@ -1,5 +1,2 @@
-import unittest
 from pydantic import BaseModel
 from typing import Union, Dict
-
-
