+++ app/database/schema.py
@@ -2,9 +2,9 @@ import enum
 from datetime import datetime
 from pydantic import BaseModel
 
-class LangEnum(enum.Enum):
-    ko = "ko" # 한글
-    en = "en" # 영어 
+# class LangEnum(enum.Enum):
+#     ko = "ko" # 한글
+#     en = "en" # 영어 
 
 class UserCreate(BaseModel):
     username: str
@@ -27,3 +27,14 @@ class MessageCreate(BaseModel):
     room_id:int
     from_id:int
     to_id:int
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+class Login(BaseModel):
+    username:str
+    password:str
+
+class AddFriend(BaseModel):
+    friend_name:str
